requirements.txt
# ----------------------
# Flask
# ----------------------
flask
requests
Flask-PyMongo
flask-admin
flask-script
markdown
pygments
python-slugify

# ----------------------
# Testing
# ----------------------
mock


# ----------------------
# Production Server
# ----------------------
gevent
gunicorn
